implement 3 configs

+++ cfd/dataset.py
@@ -11,7 +11,7 @@ from torch.utils.data import Dataset
 import torch.nn.functional as F
 
 from cfd.sensors import LHS, AroundCylinder
-from cfd.embedding import Voronoi, Mask
+from cfd.embedding import Voronoi, Mask, Vector
 
 
 class DatasetMixin:
@@ -33,15 +33,13 @@ class DatasetMixin:
         assert sensor_timeframes.shape == (self.n_chunks, len(self.init_sensor_timeframes))
         return sensor_timeframes.int()
 
-    def prepare_fullstate_timeframes(self, seed: int | None = None, init_fullstate_timeframe: int = -1) -> torch.IntTensor:
-        assert seed is not None or init_fullstate_timeframe != -1, 'must be either deterministic or random'
-        if seed is None and init_fullstate_timeframe != -1:    # deterministic
-            assert self.n_fullstate_timeframes_per_chunk == 1, (
-                f'n_fullstate_timeframes_per_chunk should be logically set to 1 when target frames are generated deterministically '
-                f'(otherwise it contains overlapping frames), '
-                f'get: {self.n_fullstate_timeframes_per_chunk}'
+    def prepare_fullstate_timeframes(self, seed: int | None = None, init_fullstate_timeframes: int | None = None) -> torch.IntTensor:
+        assert seed is not None or init_fullstate_timeframes is not None, 'must be either deterministic or random'
+        if seed is None and init_fullstate_timeframes is not None:    # deterministic
+            fullstate_timeframes: torch.Tensor = (
+                torch.arange(self.n_chunks, device='cuda').unsqueeze(1) 
+                + torch.tensor(init_fullstate_timeframes, device='cuda').unsqueeze(0)
             )
-            fullstate_timeframes: torch.Tensor = torch.arange(self.n_chunks, device='cuda').unsqueeze(1) + init_fullstate_timeframe
             assert fullstate_timeframes.shape == (self.n_chunks, self.n_fullstate_timeframes_per_chunk)
             return fullstate_timeframes
         
@@ -70,11 +68,11 @@ class CFDDataset(Dataset, DatasetMixin):
         resolution: Tuple[int, int],
         n_sensors: int,
         dropout_probabilities: List[float],
+        noise_level: float,
         sensor_generator: Literal['LHS', 'AroundCylinder'], 
-        embedding_generator: Literal['Voronoi', 'Mask'],
-        init_fullstate_timeframe: int | None,
+        embedding_generator: Literal['Voronoi', 'Mask', 'Vector'],
+        init_fullstate_timeframes: List[int] | None,
         seed: int,
-        already_preloaded: bool
     ) -> None:
         
         super().__init__()
@@ -86,10 +84,10 @@ class CFDDataset(Dataset, DatasetMixin):
         self.resolution: Tuple[int, int] = resolution
         self.n_sensors: int = n_sensors
         self.dropout_probabilities: List[float] = dropout_probabilities
-        self.init_fullstate_timeframe: int | None = init_fullstate_timeframe
+        self.noise_level: float = noise_level
+        self.init_fullstate_timeframes: List[int] | None = init_fullstate_timeframes
         self.seed: int = seed
-        self.already_preloaded: bool = already_preloaded
-        self.is_random_fullstate_frames: bool = init_fullstate_timeframe is None
+        self.is_random_fullstate_frames: bool = init_fullstate_timeframes is None
 
         self.H, self.W = resolution
         self.n_sensor_timeframes_per_chunk: int = len(init_sensor_timeframes)
@@ -103,61 +101,56 @@ class CFDDataset(Dataset, DatasetMixin):
         self.sensor_positions_dest: str = os.path.join(self.dest, 'sensor_positions')
         self.metadata_dest: str = os.path.join(self.dest, 'metadata')
 
-        if not self.already_preloaded:
-            if init_fullstate_timeframe is not None:
-                # NOTE: deterministically generate fullstate frames
-                if n_fullstate_timeframes_per_chunk != 1:
-                    raise ValueError(
-                        f'n_fullstate_timeframes_per_chunk should be logically set to 1 when sensors are generated deterministically, '
-                        f'(otherwise it contains overlapping frames), '
-                        f'get: {n_fullstate_timeframes_per_chunk}'
-                    )
-                if n_samplings_per_chunk != 1:
-                    raise ValueError(
-                        f'n_samplings_per_chunk should be logically set to 1 when sensors are generated deterministically, '
-                        f'get: {n_samplings_per_chunk}'
-                    )
-                self.n_chunks: int = self.total_timeframes_per_case - max(self.init_sensor_timeframes + [init_fullstate_timeframe])  # usable for time extrapolation
-            else:
-                self.n_chunks: int = self.total_timeframes_per_case - max(self.init_sensor_timeframes)
-            # NOTE: self.n_chunks is the number of samples in one case
-
-            if sensor_generator == 'LHS':
-                self.sensor_generator = LHS(n_sensors=n_sensors)
-                self.sensor_generator.seed = seed
-                self.sensor_generator.resolution = resolution
-                self.sensor_positions = self.sensor_generator()
-            else:
-                self.sensor_generator = AroundCylinder(n_sensors=n_sensors)
-                self.sensor_generator.seed = seed
-                self.sensor_generator.resolution = resolution
-                self.sensor_positions = self.sensor_generator(
-                    hw_meters=(0.14, 0.24), center_hw_meters=(0.07, 0.065), radius_meters=0.03,
+        if not self.is_random_fullstate_frames:
+            # NOTE: fullstate frames are deterministically generated
+            if n_fullstate_timeframes_per_chunk != len(init_fullstate_timeframes):
+                raise ValueError(
+                    f'n_fullstate_timeframes_per_chunk should be logically set to len(init_fullstate_timeframes) when sensors are generated deterministically, '
+                    f'get: n_fullstate_timeframes_per_chunk={n_fullstate_timeframes_per_chunk} and init_fullstate_timeframes={init_fullstate_timeframes}'
                 )
-            
-            assert self.sensor_positions.shape == (self.sensor_generator.n_sensors, 2)
-            if embedding_generator == 'Mask':
-                self.embedding_generator = Mask(
-                    resolution=resolution, sensor_positions=self.sensor_positions, dropout_probabilities=dropout_probabilities
-                )
-            else:
-                self.embedding_generator = Voronoi(
-                    resolution=resolution, sensor_positions=self.sensor_positions, dropout_probabilities=dropout_probabilities
+            if n_samplings_per_chunk != 1:
+                raise ValueError(
+                    f'n_samplings_per_chunk should be logically set to 1 when sensors are generated deterministically, '
+                    f'get: {n_samplings_per_chunk}'
                 )
 
-            self.sensor_timeframes = self.prepare_sensor_timeframes()
-            self.case_names: List[str] = []
-            self.sampling_ids: List[int] = []
-            self.__write2disk()
+        # NOTE: self.n_chunks is the number of samples in one case
+        self.n_chunks: int = self.total_timeframes_per_case - max(self.init_sensor_timeframes)
 
+        if sensor_generator == 'LHS':
+            self.sensor_generator = LHS(n_sensors=n_sensors)
+            self.sensor_generator.seed = seed
+            self.sensor_generator.resolution = resolution
+            self.sensor_positions = self.sensor_generator()
         else:
-            self.sensor_positions: torch.Tensor = torch.load(
-                f=os.path.join(self.sensor_positions_dest, 'pos.pt'), weights_only=True
+            self.sensor_generator = AroundCylinder(n_sensors=n_sensors)
+            self.sensor_generator.seed = seed
+            self.sensor_generator.resolution = resolution
+            self.sensor_positions = self.sensor_generator(
+                hw_meters=(0.14, 0.24), center_hw_meters=(0.07, 0.065), radius_meters=0.03,
             )
-            with open(os.path.join(self.metadata_dest, 'metadata.json'), 'r') as f:
-                records: List[Dict[str, Any]] = json.load(f)
-                self.case_names: List[str] = [record['case_name'] for record in records]
-                self.sampling_ids: List[int] = [record['sampling_id'] for record in records]
+        
+        assert self.sensor_positions.shape == (self.sensor_generator.n_sensors, 2)
+        if embedding_generator == 'Mask':
+            self.embedding_generator = Mask(
+                resolution=resolution, sensor_positions=self.sensor_positions, 
+                dropout_probabilities=dropout_probabilities, noise_level=noise_level,
+            )
+        elif embedding_generator == 'Voronoi':
+            self.embedding_generator = Voronoi(
+                resolution=resolution, sensor_positions=self.sensor_positions, 
+                dropout_probabilities=dropout_probabilities, noise_level=noise_level,
+            )
+        else:
+            self.embedding_generator = Vector(
+                resolution=resolution, sensor_positions=self.sensor_positions, 
+                dropout_probabilities=dropout_probabilities, noise_level=noise_level,
+            )
+
+        self.sensor_timeframes = self.prepare_sensor_timeframes()
+        self.case_names: List[str] = []
+        self.sampling_ids: List[int] = []
+        self.__write2disk()
 
     def __getitem__(self, idx: int) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
         prefix: str = f'_{self.case_names[idx]}_{self.sampling_ids[idx]}_'
@@ -187,24 +180,15 @@ class CFDDataset(Dataset, DatasetMixin):
 
     def __write2disk(self) -> None:
         # prepare dest directories
-        if os.path.exists(self.dest): 
-            confirmed: bool = input(
-                f'Enter "yes" to permanently DELETE the loaded dataset in {self.dest} and reload over again: '
-            ) == 'yes'
-            if confirmed:
-                shutil.rmtree(self.dest)
-            else:
-                print('Terminated')
-                sys.exit(1)
-
+        if os.path.isdir(self.dest): shutil.rmtree(self.dest)
         os.makedirs(name=self.sensor_timeframes_dest, exist_ok=True)
         os.makedirs(name=self.sensor_values_dest, exist_ok=True)
         os.makedirs(name=self.fullstate_timeframes_dest, exist_ok=True)
         os.makedirs(name=self.fullstate_values_dest, exist_ok=True)
         os.makedirs(name=self.sensor_positions_dest, exist_ok=True)
         os.makedirs(name=self.metadata_dest, exist_ok=True)
-
-        # save self.sensor_positions for self.already_preloaded = True
+        
+        # save position tensors for reference
         torch.save(obj=self.sensor_positions, f=os.path.join(self.sensor_positions_dest, 'pos.pt'))
 
         sensor_timeframes_list: List[List[int]] = []
@@ -231,7 +215,7 @@ class CFDDataset(Dataset, DatasetMixin):
                 else:
                     print('Deterministically generating fullstate frames')
                     fullstate_timeframes: torch.Tensor = self.prepare_fullstate_timeframes(
-                        init_fullstate_timeframe=self.init_fullstate_timeframe
+                        init_fullstate_timeframes=self.init_fullstate_timeframes
                     )
 
                 fullstate_data: torch.Tensor = data[fullstate_timeframes]
@@ -242,7 +226,10 @@ class CFDDataset(Dataset, DatasetMixin):
                 assert fullstate_timeframes.shape == (self.n_chunks, self.n_fullstate_timeframes_per_chunk)
                 # compute sensor data for entire space
                 sensor_data: torch.Tensor = self.embedding_generator(data=sensor_frame_data, seed=self.seed + case_id + sampling_id)
-                assert sensor_data.shape == (self.n_chunks, self.n_sensor_timeframes_per_chunk, 2, self.H, self.W)
+                if isinstance(self.embedding_generator, (Voronoi, Mask)):
+                    assert sensor_data.shape == (self.n_chunks, self.n_sensor_timeframes_per_chunk, 2, self.H, self.W)
+                else:
+                    assert sensor_data.shape == (self.n_chunks, self.n_sensor_timeframes_per_chunk, 2, self.n_sensors)
                 # Write each sample to disk
                 for idx in tqdm(range(self.n_chunks), desc=f'Case {case_id + 1} | Sampling {sampling_id + 1}: '):
                     # case name & sampling id
@@ -265,7 +252,6 @@ class CFDDataset(Dataset, DatasetMixin):
                 # manual garbage collection to optimize GPU RAM, otherwise likely lead to OutOfMemoryError
                 del fullstate_data, sensor_data
         
-        # save self.case_names for self.already_preloaded = True
         assert len(self.case_names) == len(self.sampling_ids) == len(sensor_timeframes_list) == len(fullstate_timeframes_list)
         records: List[Dict[str, Any]] = [
             {
@@ -280,23 +266,3 @@ class CFDDataset(Dataset, DatasetMixin):
             json.dump(obj=records, fp=f, indent=2)
 
 
-
-if __name__ == '__main__':
-    # NOTE: How to load precomputed dataset from disk
-    self = CFDDataset(
-        root='data/train',
-        # root='data/test',
-        init_sensor_timeframes=[0, 5, 10, 15, 20],
-        n_fullstate_timeframes_per_chunk=1,
-        n_samplings_per_chunk=1,
-        resolution=(140, 240),
-        n_sensors=32,
-        dropout_probabilities=[0.05, 0.04, 0.03, 0.02, 0.01],
-        sensor_generator='LHS',
-        embedding_generator='Voronoi',
-        init_fullstate_timeframe=None,
-        seed=1,
-        already_preloaded=False,    # set to False for first time load
-    )
-    # first sample
-    sensor_timeframe_tensor, sensor_tensor, fullstate_timeframe_tensor, fullstate_tensor, case_name, sampling_id = self[0]
